Demo: Build with RHEL Entitlements
This contains the manifests and code needed to build a container
that installs RHEL subscription content on OpenShift. It takes
advantage of the Shared Resource CSI Driver and recent improvements
to consuming entitlements in RHEL 9.3+.

Notable is the additional RBAC that needs to be granted to the Shared
Resource CSI driver for each shared object. This is an area for future
improvement.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>